requirements.txt
Django==2.2.10
djangorestframework==3.10
psycopg2-binary>=2.7.6.1
gunicorn==19.9.0
django-cors-headers==3.2.0
